# Add vectorized paths for Span<T>.Reverse