… src/libraries/System.Private.CoreLib/src/System/SpanHelpers.cs