Adds vectorized paths to `Span<T>.Reverse` for types that are supported. Falls back to previous behavior if `T` is not a value type or too big for a vector.

Compared against 65a5d0e8a3027f2b240d098f47327ba3e1f743b1.

Using [this](https://github.com/dotnet/performance/blob/36ea5c39c75ee2ca82a01a2eed22688f9eb98aed/src/benchmarks/micro/libraries/System.Memory/Span.cs#L42) microbenchmark to compare performance, modified to use more buffer sizes and types:

<details>
<summary>Microbenchmark changes</summary>

```
diff --git a/src/benchmarks/micro/libraries/System.Memory/Span.cs b/src/benchmarks/micro/libraries/System.Memory/Span.cs
index e696e141..75d28d7d 100644
--- a/src/benchmarks/micro/libraries/System.Memory/Span.cs
+++ b/src/benchmarks/micro/libraries/System.Memory/Span.cs
@@ -14,11 +14,20 @@ namespace System.Memory
     [GenericTypeArguments(typeof(byte))]
     [GenericTypeArguments(typeof(char))]
     [GenericTypeArguments(typeof(int))]
+    [GenericTypeArguments(typeof(long))]
+    [GenericTypeArguments(typeof(float))]
+    [GenericTypeArguments(typeof(double))]
     [BenchmarkCategory(Categories.Runtime, Categories.Libraries, Categories.Span)]
     public class Span<T>
         where T : struct, IComparable<T>, IEquatable<T>
     {
-        [Params(Utils.DefaultCollectionSize)]
+        [Params(
+            8 /* No vectorization */,
+            34 /* SSSE3 with leftover */,
+            68 /* AVX2 path with leftover bytes */,
+            Utils.DefaultCollectionSize,
+            Utils.DefaultCollectionSize * 2
+            )]
         public int Size;

         private T[] _array, _same, _emptyWithSingleValue;
```

</details>

Performance results:

```
$ py .\scripts\benchmarks_ci.py -c Release -f net7.0 --filter *Span*Reverse* --corerun C:\Users\acovingt\source\repos\runtime-master\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts C:\Users\acovingt\Documents\vectorize-span-reverse-base
$ py .\scripts\benchmarks_ci.py -c Release -f net7.0 --filter *Span*Reverse* --corerun C:\Users\acovingt\source\repos\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts C:\Users\acovingt\Documents\vectorize-span-reverse-diff
$ cd .\src\tools\ResultsComparer\ 
$ dotnet run -- --base C:\Users\acovingt\Documents\vectorize-span-reverse-base\ --diff C:\Users\acovingt\Documents\vectorize-span-reverse-diff\ --threshold 3% --noise 5ns
summary:
better: 23, geomean: 3.946
total diff: 23

No Slower results for the provided threshold = 3% and noise filter = 5ns.

| Faster                                         | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Memory.Span<Byte>.Reverse(Size: 1024)   |     33.45 |           616.32 |            18.42 |         |
| System.Memory.Span<Byte>.Reverse(Size: 512)    |     28.16 |           306.00 |            10.87 |         |
| System.Memory.Span<Char>.Reverse(Size: 512)    |     16.26 |           300.26 |            18.47 |         |
| System.Memory.Span<Char>.Reverse(Size: 1024)   |     14.25 |           609.56 |            42.77 |         |
| System.Memory.Span<Byte>.Reverse(Size: 68)     |      5.75 |            33.59 |             5.84 |         |
| System.Memory.Span<Char>.Reverse(Size: 68)     |      5.54 |            39.56 |             7.14 | bimodal |
| System.Memory.Span<Single>.Reverse(Size: 512)  |      4.57 |           152.19 |            33.33 |         |
| System.Memory.Span<Char>.Reverse(Size: 34)     |      4.13 |            21.56 |             5.22 | bimodal |
| System.Memory.Span<Single>.Reverse(Size: 1024) |      3.93 |           303.01 |            77.20 |         |
| System.Memory.Span<Int32>.Reverse(Size: 1024)  |      3.88 |           307.61 |            79.20 |         |
| System.Memory.Span<Int32>.Reverse(Size: 512)   |      3.67 |           154.75 |            42.20 |         |
| System.Memory.Span<Byte>.Reverse(Size: 34)     |      3.03 |            15.56 |             5.14 | several?|
| System.Memory.Span<Int32>.Reverse(Size: 68)    |      2.59 |            23.28 |             8.98 |         |
| System.Memory.Span<Single>.Reverse(Size: 68)   |      2.45 |            19.77 |             8.06 |         |
| System.Memory.Span<Single>.Reverse(Size: 34)   |      2.21 |            12.41 |             5.62 | several?|
| System.Memory.Span<Int64>.Reverse(Size: 1024)  |      2.01 |           302.07 |           150.24 |         |
| System.Memory.Span<Int32>.Reverse(Size: 34)    |      1.99 |            15.11 |             7.61 |         |
| System.Memory.Span<Int64>.Reverse(Size: 512)   |      1.98 |           153.95 |            77.64 |         |
| System.Memory.Span<Double>.Reverse(Size: 512)  |      1.97 |           152.21 |            77.18 |         |
| System.Memory.Span<Double>.Reverse(Size: 1024) |      1.93 |           301.68 |           156.33 |         |
| System.Memory.Span<Int64>.Reverse(Size: 68)    |      1.91 |            23.54 |            12.35 |         |
| System.Memory.Span<Int64>.Reverse(Size: 34)    |      1.76 |            14.90 |             8.49 |         |
| System.Memory.Span<Double>.Reverse(Size: 68)   |      1.64 |            19.80 |            12.08 |         |
```

Please let me know if I can provide any other information. Thanks!