@stephentoub 2022-01-28
Thanks for sharing the perf tests.  The results only show down to a size of 32 elements, and all of them show improvements.  Is there a smaller size at which this is actually a regression?