@alexcovington 2022-01-28
> Thanks for sharing the perf tests. The results only show down to a size of 32 elements, and all of them show improvements. Is there a smaller size at which this is actually a regression?

@stephentoub I didn't notice that the table didn't include the results for 8 byte spans. I changed my filter to be a little more inclusive when comparing results:

```
PS C:\Users\acovingt\source\repos\performance\src\tools\ResultsComparer> dotnet run -- --base C:\Users\acovingt\Documents\vectorize-span-reverse-base\ --diff C:\Users\acovingt\Documents\vectorize-span-reverse-diff\ --threshold 1% --noise 1ns
summary:
better: 24, geomean: 3.806
worse: 4, geomean: 1.463
total diff: 28

| Slower                                      | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Memory.Span<Double>.Reverse(Size: 8) |      1.69 |             4.20 |             7.08 |         |
| System.Memory.Span<Single>.Reverse(Size: 8) |      1.55 |             4.21 |             6.53 |         |
| System.Memory.Span<Int32>.Reverse(Size: 8)  |      1.34 |             5.02 |             6.74 |         |
| System.Memory.Span<Byte>.Reverse(Size: 8)   |      1.30 |             5.24 |             6.82 |         |

| Faster                                         | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Memory.Span<Byte>.Reverse(Size: 1024)   |     33.45 |           616.32 |            18.42 |         |
| System.Memory.Span<Byte>.Reverse(Size: 512)    |     28.16 |           306.00 |            10.87 |         |
| System.Memory.Span<Char>.Reverse(Size: 512)    |     16.26 |           300.26 |            18.47 |         |
| System.Memory.Span<Char>.Reverse(Size: 1024)   |     14.25 |           609.56 |            42.77 |         |
| System.Memory.Span<Byte>.Reverse(Size: 68)     |      5.75 |            33.59 |             5.84 |         |
| System.Memory.Span<Char>.Reverse(Size: 68)     |      5.54 |            39.56 |             7.14 | bimodal |
| System.Memory.Span<Single>.Reverse(Size: 512)  |      4.57 |           152.19 |            33.33 |         |
| System.Memory.Span<Char>.Reverse(Size: 34)     |      4.13 |            21.56 |             5.22 | bimodal |
| System.Memory.Span<Single>.Reverse(Size: 1024) |      3.93 |           303.01 |            77.20 |         |
| System.Memory.Span<Int32>.Reverse(Size: 1024)  |      3.88 |           307.61 |            79.20 |         |
| System.Memory.Span<Int32>.Reverse(Size: 512)   |      3.67 |           154.75 |            42.20 |         |
| System.Memory.Span<Byte>.Reverse(Size: 34)     |      3.03 |            15.56 |             5.14 | several?|
| System.Memory.Span<Int32>.Reverse(Size: 68)    |      2.59 |            23.28 |             8.98 |         |
| System.Memory.Span<Single>.Reverse(Size: 68)   |      2.45 |            19.77 |             8.06 |         |
| System.Memory.Span<Single>.Reverse(Size: 34)   |      2.21 |            12.41 |             5.62 | several?|
| System.Memory.Span<Int64>.Reverse(Size: 1024)  |      2.01 |           302.07 |           150.24 |         |
| System.Memory.Span<Int32>.Reverse(Size: 34)    |      1.99 |            15.11 |             7.61 |         |
| System.Memory.Span<Int64>.Reverse(Size: 512)   |      1.98 |           153.95 |            77.64 |         |
| System.Memory.Span<Double>.Reverse(Size: 512)  |      1.97 |           152.21 |            77.18 |         |
| System.Memory.Span<Double>.Reverse(Size: 1024) |      1.93 |           301.68 |           156.33 |         |
| System.Memory.Span<Int64>.Reverse(Size: 68)    |      1.91 |            23.54 |            12.35 |         |
| System.Memory.Span<Int64>.Reverse(Size: 34)    |      1.76 |            14.90 |             8.49 |         |
| System.Memory.Span<Double>.Reverse(Size: 34)   |      1.66 |            12.35 |             7.44 |         |
| System.Memory.Span<Double>.Reverse(Size: 68)   |      1.64 |            19.80 |            12.08 |         |
```

There's a 2-3ns regression for the really small spans possibly due to the overhead of the extra conditional checks, or it could just be noise. 

Let me know if you'd like more analysis on it.