@alexcovington 2022-01-28
> `Array.Reverse<T>` has its own almost identical implementation that won't benefit from these improvements. Can we change Array to delegate to the same underlying implementation being added here so that both arrays and spans benefit equally?
> 
> https://github.com/dotnet/runtime/blob/953fd351ae1c5c8c11ba865b3613b63f8f774624/src/libraries/System.Private.CoreLib/src/System/Array.cs#L1610-L1641

I didn't notice this, but you're right that Array can also benefit from this. I'll try adding the paths there as well and send an update.