@alexcovington 2022-03-21
@saucecontrol I've finished benchmarking using `DOTNET_EnableAVX2=0` and `DOTNET_EnableAVX2=1` to measure the difference between the AVX2 and SSSE3/SSE2 path, and it looks like there is a performance improvement with the AVX2 path even with the extra permute step.

<details>

<summary>Benchmarking steps</summary>

```powershell
$env:DOTNET_EnableAVX2=1
py .\scripts\benchmarks_ci.py -c Release -f net7.0 --filter *Span*Reverse* --corerun C:\Users\acovingt\source\repos\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts D:\workspace\20220318\vectorize-span-reverse-bench\avx2-enabled
$env:DOTNET_EnableAVX2=0
py .\scripts\benchmarks_ci.py -c Release -f net7.0 --filter *Span*Reverse* --corerun C:\Users\acovingt\source\repos\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts D:\workspace\20220318\vectorize-span-reverse-bench\avx2-disabled
cd .\src\tools\ResultsComparer\
dotnet run -- --base D:\workspace\20220318\vectorize-span-reverse-bench\avx2-disabled\ --diff D:\workspace\20220318\vectorize-span-reverse-bench\avx2-enabled\ --threshold 1% --noise 1ns --full-id
```
</details> 

Using `DOTNET_EnableAVX2=0` as the baseline and `DOTNET_EnableAVX2=1` as the difference, here is a chart summarizing the same benchmark:

```
PS C:\Users\acovingt\source\repos\performance\src\tools\ResultsComparer> dotnet run -- --base D:\workspace\20220318\vectorize-span-reverse-bench\avx2-disabled\ --diff D:\workspace\20220318\vectorize-span-reverse-bench\avx2-enabled\ --threshold 1% --noise 1ns --full-id
summary:
better: 43, geomean: 1.469
total diff: 43

No Slower results for the provided threshold = 1% and noise filter = 1ns.

| Faster                                              | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Memory.Span<Single>.Reverse(Size: 1024)      |      1.70 |           122.03 |            71.75 |         |
| System.Memory.Span<Int64>.ReverseArray(Size: 34)    |      1.67 |            10.29 |             6.15 |         |
| System.Memory.Span<Byte>.ReverseArray(Size: 1024)   |      1.67 |            31.33 |            18.78 |         |
| System.Memory.Span<Int64>.ReverseArray(Size: 512)   |      1.65 |           120.61 |            73.26 |         |
| System.Memory.Span<Single>.ReverseArray(Size: 68)   |      1.65 |            10.64 |             6.47 |         |
| System.Memory.Span<Char>.ReverseArray(Size: 512)    |      1.63 |            27.76 |            17.03 |         |
| System.Memory.Span<Single>.Reverse(Size: 512)       |      1.62 |            59.32 |            36.61 |         |
| System.Memory.Span<Double>.ReverseArray(Size: 68)   |      1.60 |            17.23 |            10.75 |         |
| System.Memory.Span<Int32>.Reverse(Size: 512)        |      1.60 |            59.31 |            37.13 |         |
| System.Memory.Span<Int64>.ReverseArray(Size: 68)    |      1.59 |            17.06 |            10.73 | bimodal |
| System.Memory.Span<Int64>.ReverseArray(Size: 8)     |      1.56 |             4.07 |             2.61 |         |
| System.Memory.Span<Double>.ReverseArray(Size: 512)  |      1.56 |           120.45 |            77.34 |         |
| System.Memory.Span<Single>.ReverseArray(Size: 1024) |      1.56 |           120.25 |            77.26 |         |
| System.Memory.Span<Double>.Reverse(Size: 512)       |      1.54 |           122.38 |            79.67 |         |
| System.Memory.Span<Int64>.Reverse(Size: 512)        |      1.54 |           121.86 |            79.34 |         |
| System.Memory.Span<Byte>.ReverseArray(Size: 512)    |      1.53 |            17.49 |            11.43 |         |
| System.Memory.Span<Double>.Reverse(Size: 1024)      |      1.52 |           244.01 |           160.74 |         |
| System.Memory.Span<Int32>.Reverse(Size: 1024)       |      1.50 |           120.50 |            80.11 |         |
| System.Memory.Span<Byte>.Reverse(Size: 1024)        |      1.50 |            27.74 |            18.47 |         |
| System.Memory.Span<Char>.Reverse(Size: 512)         |      1.50 |            31.03 |            20.73 |         |
| System.Memory.Span<Double>.ReverseArray(Size: 1024) |      1.49 |           233.93 |           156.49 |         |
| System.Memory.Span<Int32>.ReverseArray(Size: 68)    |      1.49 |            10.06 |             6.73 |         |
| System.Memory.Span<Int32>.ReverseArray(Size: 1024)  |      1.48 |           118.63 |            79.94 |         |
| System.Memory.Span<Byte>.Reverse(Size: 512)         |      1.48 |            16.50 |            11.13 |         |
| System.Memory.Span<Char>.ReverseArray(Size: 1024)   |      1.46 |            60.86 |            41.65 |         |
| System.Memory.Span<Double>.ReverseArray(Size: 34)   |      1.46 |             9.15 |             6.26 |         |
| System.Memory.Span<Int64>.ReverseArray(Size: 1024)  |      1.43 |           231.05 |           161.07 |         |
| System.Memory.Span<Single>.Reverse(Size: 68)        |      1.43 |            11.95 |             8.36 |         |
| System.Memory.Span<Single>.ReverseArray(Size: 512)  |      1.42 |            58.29 |            40.92 |         |
| System.Memory.Span<Int32>.ReverseArray(Size: 34)    |      1.42 |             5.85 |             4.11 |         |
| System.Memory.Span<Int32>.ReverseArray(Size: 512)   |      1.42 |            57.23 |            40.28 |         |
| System.Memory.Span<Single>.ReverseArray(Size: 34)   |      1.42 |             5.90 |             4.16 |         |
| System.Memory.Span<Int64>.Reverse(Size: 34)         |      1.41 |            11.92 |             8.43 |         |
| System.Memory.Span<Int32>.Reverse(Size: 34)         |      1.40 |             9.07 |             6.48 | bimodal |
| System.Memory.Span<Int32>.Reverse(Size: 68)         |      1.39 |            12.67 |             9.11 |         |
| System.Memory.Span<Char>.Reverse(Size: 1024)        |      1.34 |            57.71 |            43.19 |         |
| System.Memory.Span<Int64>.Reverse(Size: 1024)       |      1.33 |           242.14 |           182.51 | bimodal |
| System.Memory.Span<Double>.Reverse(Size: 68)        |      1.31 |            18.42 |            14.11 |         |
| System.Memory.Span<Char>.Reverse(Size: 68)          |      1.30 |            10.68 |             8.21 |         |
| System.Memory.Span<Single>.Reverse(Size: 34)        |      1.24 |             7.10 |             5.73 |         |
| System.Memory.Span<Double>.Reverse(Size: 34)        |      1.23 |            11.20 |             9.11 |         |
| System.Memory.Span<Char>.ReverseArray(Size: 68)     |      1.23 |             6.22 |             5.06 | bimodal |
| System.Memory.Span<Int64>.Reverse(Size: 68)         |      1.17 |            18.41 |            15.77 | bimodal |
```

Please let me know if there is anything else I can check. Thanks for the all the feedback!